+++ src/main/java/dankook/cs/aj24/domain/jwt/JwtUtils.java
@@ -1,6 +1,6 @@
 package dankook.cs.aj24.domain.jwt;
 
-import dankook.cs.aj24.common.util.UserRole;
+import dankook.cs.aj24.domain.user.UserRole;
 import dankook.cs.aj24.domain.oauth.PrincipalDetail;
 import dankook.cs.aj24.domain.user.UserDocument;
 import io.jsonwebtoken.ExpiredJwtException;
