+++ src/main/java/dankook/cs/aj24/domain/user/UserRole.java
@@ -1,4 +1,4 @@
-package dankook.cs.aj24.common.util;
+package dankook.cs.aj24.domain.user;
 
 import lombok.Getter;
